backEnd/book_manger/requirement.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-rest-knox==5.0.1
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
pymongo==4.8.0
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
